tools/testing/requirements_openai.txt
# Requirements for OpenAI API Testing
# Install with: pip install -r requirements_openai.txt

# OpenAI API
openai>=1.0.0

# Data analysis and visualization
pandas>=2.0.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Async support (should already be installed)
asyncio
aiohttp>=3.8.0

# Core dependencies (from main project)
numpy>=1.24.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0

# Testing utilities
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For better progress bars
tqdm>=4.65.0

# Optional: For colored terminal output
colorama>=0.4.6
rich>=13.0.0
